# Conflicts:
#	android/PieShare/app/src/main/java/org/pieShare/android/userTool/UserTools.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/task/commandTasks/loginTask/LoginTask.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/task/localTasks/fileEventTask/ALocalFileEventTask.java
#	core/pieShareApp/src/main/java/org/pieShare/pieShareApp/task/localTasks/fileEventTask/LocalFileCreatedTask.java

+++ pieSpring/src/main/java/org/pieshare/piespring/service/database/DatabaseService.java
@@ -69,6 +69,10 @@ public class DatabaseService implements IDatabaseService {
     @Override
     public ArrayList<PieUser> findAllPieUser() {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(PieUserEntity.class);
+		
+		if(!em.isOpen()) {
+			return null;
+		}
 
         Query query = em.createQuery(String.format("SELECT e FROM %s e", PieUserEntity.class.getSimpleName()), PieUserEntity.class);
         PieUser user = null;
@@ -95,14 +99,26 @@ public class DatabaseService implements IDatabaseService {
     public void removePieUser(PieUser user) {
         PieUserEntity ent;
         ent = (PieUserEntity) modelEntityConverterService.convertToEntity(user);
-        remove(ent);
+        EntityManager em = pieDatabaseManagerFactory.getEntityManger(ent.getClass());
+		if(!em.isOpen()) {
+			return;
+		}
+        em.getTransaction().begin();
+        em.remove(em.merge(ent));
+        em.getTransaction().commit();
     }
 
     @Override
     public void mergePieUser(PieUser user) {
         PieUserEntity entity;
         entity = (PieUserEntity) modelEntityConverterService.convertToEntity(user);
-        merge(entity);
+        EntityManager em = pieDatabaseManagerFactory.getEntityManger(entity.getClass());
+		if(!em.isOpen()) {
+			return;
+		}
+        em.getTransaction().begin();
+        em.merge(entity);
+        em.getTransaction().commit();
     }
 
     @Override
@@ -122,6 +138,9 @@ public class DatabaseService implements IDatabaseService {
     @Override
     public ArrayList<IFilter> findAllFilters() {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(FilterEntity.class);
+		if(!em.isOpen()) {
+			return null;
+		}
         Query query = em.createQuery(String.format("SELECT e FROM %s e", FilterEntity.class.getSimpleName()), FilterEntity.class);
         ArrayList<IFilter> list = new ArrayList<>();
 
@@ -138,12 +157,15 @@ public class DatabaseService implements IDatabaseService {
 
     public void persist(PieFile file) {
         PieFileEntity entity;
-        entity = (PieFileEntity)this.modelEntityConverterService.convertToEntity(file);
+        entity = (PieFileEntity) this.modelEntityConverterService.convertToEntity(file);
         this.persist(entity);
     }
 
     private void persist(IBaseEntity entity) {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(entity.getClass());
+		if(!em.isOpen()) {
+			return;
+		}
         em.getTransaction().begin();
         em.persist(entity);
         em.getTransaction().commit();
@@ -151,6 +173,9 @@ public class DatabaseService implements IDatabaseService {
 
     private synchronized void merge(IBaseEntity entity) {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(entity.getClass());
+		if(!em.isOpen()) {
+			return;
+		}
         em.getTransaction().begin();
         em.merge(entity);
         em.getTransaction().commit();
@@ -158,6 +183,9 @@ public class DatabaseService implements IDatabaseService {
 
     private void remove(IBaseEntity entity) {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(entity.getClass());
+		if(!em.isOpen()) {
+			return;
+		}
         em.getTransaction().begin();
         em.remove(entity);
         em.getTransaction().commit();
@@ -166,7 +194,10 @@ public class DatabaseService implements IDatabaseService {
     @Override
     public PieFile findPieFile(PieFile file) {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(PieFileEntity.class);
-		PieFileEntity historyFileEntity = em.find(PieFileEntity.class, file.getRelativePath());
+		if(!em.isOpen()) {
+			return null;
+		}
+        PieFileEntity historyFileEntity = em.find(PieFileEntity.class, file.getRelativePath());
         return this.modelEntityConverterService.convertFromEntity(historyFileEntity);
     }
 
@@ -185,6 +216,9 @@ public class DatabaseService implements IDatabaseService {
     @Override
     public List<PieFile> findAllUnsyncedPieFiles() {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(PieFileEntity.class);
+		if(!em.isOpen()) {
+			return null;
+		}
         String sqlQuery = String.format("SELECT f FROM %s f WHERE f.synched=TRUE", PieFileEntity.class.getSimpleName());
         TypedQuery<PieFileEntity> query = em.createQuery(sqlQuery, PieFileEntity.class);
 
@@ -202,6 +236,9 @@ public class DatabaseService implements IDatabaseService {
     @Override
     public void resetAllPieFileSynchedFlags() {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(PieFileEntity.class);
+		if(!em.isOpen()) {
+			return;
+		}
         String sqlQuery = String.format("UPDATE %s SET synched=TRUE", PieFileEntity.class.getSimpleName());
         TypedQuery<PieFileEntity> query = em.createQuery(sqlQuery, PieFileEntity.class);
         em.getTransaction().begin();
@@ -212,6 +249,9 @@ public class DatabaseService implements IDatabaseService {
     @Override
     public List<PieFile> findAllPieFiles() {
         EntityManager em = pieDatabaseManagerFactory.getEntityManger(PieFileEntity.class);
+		if(!em.isOpen()) {
+			return null;
+		}
         String sqlQuery = String.format("SELECT f FROM %s f", PieFileEntity.class.getSimpleName());
         TypedQuery<PieFileEntity> query = em.createQuery(sqlQuery, PieFileEntity.class);
